Switch to Rouge highlighter
Jekyll has made rouge the default.

The Fortran lexer in pygments wasn't lexing exponents correctly.  I
submitted a pull request, which has been merged, but is not in a release
version at

    https://bitbucket.org/birkenfeld/pygments-main/pull-requests/656/fortran-exponent-letter/diff

The Fortran lexer in rouge also wasn't lexing exponents correctly.  I
submitted a pull request, which has been merged into master.  As of
Jekyll 3.2, we can only use versions ~> 1.7 of rouge.  So for now, I am
making a gem of rouge v1.11.1 with my commit cherry-picked onto it.